Nahid

+++ app/Http/Controllers/admin/RolePermissionController.php
@@ -16,9 +16,50 @@ class RolePermissionController extends Controller
 {
     public function roleIndex(){
         if(auth()->user()->hasPermission('admin role index')){
-            $roles = Role::latest()->paginate(20);
-            $permissions = Permission::latest()->where('status',1)->get();
-            return view('admin.role-permission.role-index',compact('roles','permissions'));
+            $roles = Role::latest()->paginate(500);
+            $permissionsGroup = Permission::where('status',1)->get()
+                ->groupBy(function ($permission) {
+                    if (str_contains($permission->name, 'core value ')) {
+                        return 'Core Value ';
+                    } elseif (str_contains($permission->name, 'concern')) {
+                        return 'Concern';
+                    } elseif (str_contains($permission->name, 'product')) {
+                        return 'Product';
+                    } elseif (str_contains($permission->name, 'article')) {
+                        return 'Article';
+                    } elseif (str_contains($permission->name, 'post page')) {
+                        return 'Post Page';
+                    } elseif (str_contains($permission->name, 'shop')) {
+                        return 'Shop';
+                    } elseif (str_contains($permission->name, 'gallery')) {
+                        return 'Gallery ';
+                    } elseif (str_contains($permission->name, 'role')) {
+                        return 'Role';
+                    } elseif (str_contains($permission->name, 'permission')) {
+                        return 'Permission';
+                    } elseif (str_contains($permission->name, 'setting')) {
+                        return 'Settings';
+                    } elseif (str_contains($permission->name, 'slider')) {
+                        return 'Slider';
+                    } elseif (str_contains($permission->name, 'about')) {
+                        return 'About Us';
+                    } elseif (str_contains($permission->name, 'topbar')) {
+                        return 'Top Bar';
+                    } elseif (str_contains($permission->name, 'cms')) {
+                        return 'Website CMS';
+                    } elseif (str_contains($permission->name, 'pass')) {
+                        return 'Password';
+                    } elseif (str_contains($permission->name, 'user')) {
+                        return 'User';
+                    } elseif (str_contains($permission->name, 'block email')) {
+                        return 'Block Email';
+                    } elseif (str_contains($permission->name, 'social')) {
+                        return 'Social Link';
+                    } elseif (str_contains($permission->name, '')) {
+                        return 'Missions';
+                    }
+                });
+            return view('admin.role-permission.role-index',compact('roles','permissionsGroup'));
         }
         else{
             toastr()->error('You Have No Permission.');
@@ -42,12 +83,17 @@ class RolePermissionController extends Controller
                 $role->permission_ids = json_encode($request->permission_ids);
                 $role->status = $request->status;
                 $role->save();
-
-                foreach ($request->permission_ids as $permission){
-                    $permit = new RolePermission();
-                    $permit->role_id = $role->id;
-                    $permit->permission_id = $permission;
-                    $permit->save();
+                if (!empty($request->permission_ids)){
+                    foreach ($request->permission_ids as $permission){
+                        $permit = new RolePermission();
+                        $permit->role_id = $role->id;
+                        $permit->permission_id = $permission;
+                        $permit->save();
+                    }
+                }
+                else{
+                    toastr()->error('Role Create Success But No Permission Selected.');
+                    return back();
                 }
                 toastr()->success('Role Create Success.');
                 return back();
@@ -68,6 +114,7 @@ class RolePermissionController extends Controller
             try{
                 $validate = Validator::make($request->all(),[
                     'name' => 'required',
+                    'permission_ids' => 'required',
                 ]);
                 if($validate->fails()){
                     toastr()->error($validate->messages());
@@ -78,19 +125,24 @@ class RolePermissionController extends Controller
                 $role->permission_ids = json_encode($request->permission_ids);
                 $role->status = $request->status;
                 $role->save();
-
-                $rolePermissions = RolePermission::where('role_id',$id)->whereNotIn('permission_id',$request->permission_ids)->pluck('id')->toArray();
-                RolePermission::destroy($rolePermissions);
-
-                foreach ($request->permission_ids as $permission){
-                    $permit = RolePermission::where('role_id',$id)->where('permission_id',$permission)->first();
-                    if (!$permit){
-                        $permit = new RolePermission();
-                        $permit->role_id = $role->id;
-                        $permit->permission_id = $permission;
-                        $permit->save();
+                if (!empty($request->permission_ids)){
+                    $rolePermissions = RolePermission::where('role_id',$id)->whereNotIn('permission_id',$request->permission_ids)->pluck('id')->toArray();
+                    RolePermission::destroy($rolePermissions);
+                    foreach ($request->permission_ids as $permission){
+                        $permit = RolePermission::where('role_id',$id)->where('permission_id',$permission)->first();
+                        if (!$permit){
+                            $permit = new RolePermission();
+                            $permit->role_id = $role->id;
+                            $permit->permission_id = $permission;
+                            $permit->save();
+                        }
                     }
                 }
+                else{
+                    toastr()->error('No Permission Selected.');
+                    return back();
+                }
+
                 toastr()->success('Role Update Success.');
                 return back();
             }
